fixed post request and delete hooks in component folder
